Added solution for day 14
To be honest, part 2 was kind of stupid. There's no indication in the problem about what "arranged into a Christmas tree" means, and given how many iterations it took, manually looking through all of the outputs is dumb. Someone pointed out that "robots that make a tree are unlikely to overlap" (which is how I solved it), but there's 0 indication of this in the problem.